This commit removes state mgmt and the global enable.
If any system is disabled, it should either be due to a disable
being sent from the user, an override, or a fault. This means that
there is no need for us to maintain all systems' states in the driver.
We can just continue to spam the most recent command and only modify
it if we receive an override_active or fault_active flag on the global
command. If either of those are true, we immediately disable all
systems.